vayu: audio: Add built-in/Back mic
Fix for vayu built-in mics missing in the primary_input.

It's recomennded to add (pulse custom.pa):
  set-source-port source.primary_input input-builtin_mic

Source:
https://github.com/SailfishOS-vayu/android_device_xiaomi_sm8150-common/commit/1706c8fbe228e598db8d8c56cef4c2f11ef966af

Before this fix, the built-in mics were only exported to the
surround_sound source.

But pulseaudio is not loading the surround_sound source,
but only the primary_input.

This is causing a missing mics problem to the applications
since they looks at the primary_input, and there are no mics exported.

This commit adds the built-in mics to the primary_input source,
but keeps the original exports in the surround_sound source
to avoid audio breaks.

The original commit was adapted to the Droidian overlay mechanism, and
some information was added to the commit message body.

Signed-off-by: berbascum <berbascum@ticv.cat>